Clarify priorities to be relative not absolute
Article 2.2.2 states a Lead Link may define priorities for the Circle. I would suggest to specify a Lead Link can define *relative* priorities for the Circle, not *absolute* priorities, similar to the use in article 4.1.1 b). See http://fabiopereira.me/blog/2009/06/29/principle-of-relative-priority/ for a brief discussion of the difference between relative and absolute priorities.