Channel for out-of-meeting governance
If the intent of Constitution 4.1 is just to simplify language, and not to evolve the rules, then this would be a more accurate evolution. As a side note, I have experienced a situation as Secretary for the Ecosystem circle where I wanted to use Slack for governance proposals, but I couldn't legitimately consider it a "typically-used channel" given how little use was made of it by most Provider Cross Links. So I proposed a policy to make it a valid channel to propose governance. I believe this threshold of "typically-used channel" rightfully restricted me so as to not impose my technological choice to the rest of the circle. So I pitch for keeping that threshold and not giving Secretary the authority to define those channels on its own. 

Also, the way it reads on this version (before my suggested change) could be interpreted as saying that any written communication channel must be explicitly approved a priori by the Secretary ("I approve email"), which would be impractical.