checklist metrics swap
I had questions on the order of the tactical meeting. and that it felt unnatural to do checklists and project updates separate.
with this new order we lookk first at the circle as a whole with the metrics on how we are doing. and then with checklists and project updates dive into the actual work of the circle. It also helps that metrics are defined by the LL and checklists/projects can be added by any role themselves.