Allowed representing tensions in tactical via IA
Removed the restriction that you must fill a Role to process tensions
for it in a tactical meeting, to allow processing for Individual
Action; however, kept the recently-added restriction that it must still
be to serve a Role invited to the meeting (i.e. not a personal tension,
nor one for an uninvited role).