Removed the defined term "Circle Member"
After recent changes, Circle Member was no longer really used, at least
not in a meaningful way that wasn’t easily replaced with another
distinction.  This seems super weird, but that’s where I got
accidentally arrived by following the tensions and solving other
issues.  So it seems the distinction is no longer useful, so I removed
it.  I’ve separately captured an action to consider re-introducing it
as a distinction to mean all Role Representatives invited to govern a
Circle (i.e. what Core Circle Member used to mean), to simplify some of
the places where a more clunky phrase is used in Article 3.