Update Holacracy-Constitution.md
In my experience, it's far more common for an objection to be invalidated because it doesn't fit one of the objector's roles, rather than because it's predictive. It is also a more important learning point ("Is the tension felt in a role you fill?") for early practice. However, the current version of the Constitution sequences the criteria in such a way that many Facilitators ask the predictive question (and all of its nuanced explanations) when if the Facilitator had just asked the role-based question, it would have likely been invalidated. So, I'm proposing a simple re-ordering of the criteria to make it easier for Facilitators to apply the test questions