Update Holacracy-Constitution.md
Truth is, people with laptop add points even during processing points, with GlassFrog interface allowing this. I wonder the why of this? I know it's because Human can have “hot reactions”, and if you let a few minutes, just like a soufflé, it gets down and they no longer feel the tension (example of an objection, which the persons say he/she'll bring it to the agenda to process it, and 95% of the time he/she doesn't). The concrete case of why it have been added in the first place would be interesting, because “it may disrupt a bit the point which is being processed”, well, no true, as long as you have the possibility to speak you may for example ask the Secretary to see GlassFrog, so why not add a point to the agenda? I'd really love the tension at the start.
I made a proposal, re-taking for the ending part the Clarifying Questions paragraph.