First cut of shifting spending authority from Circle Lead to announcing intent
This may resolve #142, however I’m concerned that “whatever Circle has
the right to spend the money or assets you intend to spend” in this new
version of §2.1.3 may be confusing; it’s clarified in §4.2 that the
Anchor Circle holds this authority, but that’s well after 2.1.3.  Is
that clear enough, or should the definition of Anchor Circle get moved
to Article 1, or something else?