Resolved #280
Also removed the direct capacity for a Lead Link to invite a guest into governance to help process a tension, as they can now effectively achieve the same thing by making someone else a Role Supporter of their Role in the Super-Circle (and thus a Circle Lead in the sub-circle).  It's also dubious whether this feature was ever really needed or used anyway.