Reorganized & Split Apart Articles
Work towards #339.  Now, the Preamble and Article 1 define a core basic package of Holacracy's structure, and define all key terms.  Every other Article stands-alone as an addition on top of that, with each adding a single cohesive function, which can be adopted one at a time and in any order (although the current ordering seems the most likely default learning path).