Clarified interpretation clauses
There was potential ambiguity in the connection of the two different interpretation sections, and potential holes if either article were adopted without the other.  This change makes the connection more clear if both articles are adopted, and also allows more clear usage with either article alone along with the envisioned stub.